cpufreq: cpu-boost: let the big ass cluster be boosted on touch
big cluster governor tunables have been quite changed for this release
so things are tamed while letting the little cluster do most of the hard job
(and being quite good at it). So with the input_boost_ms being scaled down to 40ms
(stock is 1500ms - a hopping second and a half of continuous boosting - yikes!)
we should be fine with boosting the big one on touch now.

Change-Id: Iaef7ba151b1264a221fc94b30d749d1d65ed327f
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>